Rebase

+++ chips/imxrt1050/src/lib.rs
@@ -249,11 +249,11 @@ pub static IRQS: [unsafe extern "C" fn(); 160] = [
 ];
 
 extern "C" {
-    static mut _szero: u32;
-    static mut _ezero: u32;
-    static mut _etext: u32;
-    static mut _srelocate: u32;
-    static mut _erelocate: u32;
+    static mut _szero: usize;
+    static mut _ezero: usize;
+    static mut _etext: usize;
+    static mut _srelocate: usize;
+    static mut _erelocate: usize;
 }
 
 pub unsafe fn init() {
